% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irfBand.R
\name{irfBand}
\alias{irfBand}
\title{VAR impulse responses with error bands}
\usage{
irfBand(
  pdout,
  smat = NULL,
  nstep = 40,
  order = NULL,
  pctiles = c(5, 16, 50, 84, 95),
  whichv = NULL,
  whichs = NULL,
  main = "IRF's with bands",
  file = "IRFwBandPlot.pdf"
)
}
\arguments{
\item{pdout}{Output from \code{\link[=postdraw`]{postdraw`()}} or \code{\link[=SVARpostdraw]{SVARpostdraw()}}.}

\item{smat}{With \code{postdraw} (from \code{\link[=rfvar]{rfvar()}} output, if NULL), the transpose of
\code{pdout$smat[ , , draw]} is used as initial shocks.  Otherwise,
this array is used, not transposed, and so must have variables
in its rows and shocks in its columns, as \code{\link[=impulsdtrf]{impulsdtrf()}}
requires. With SVAR output, the inverse of \code{A0} is used as
initial shocks when \code{smat} is NULL.  \code{smat}
Can be a single matrix, used repeatedly.}

\item{nstep}{Number of steps ahead to calculate responses.}

\item{order}{If non-null, use triangular orthogonalization with this ordering of shocks.}

\item{pctiles}{At what percentiles to draw bands. (vector of odd length)}

\item{whichv}{logical or numeric vector picking which variables' responses are calculated}

\item{whichs}{logical or numeric vector picking which shock effects are calculated}

\item{main}{Character string giving plot title.}

\item{file}{Character string giving name of pdf file to be written.}
}
\value{
nvar x nshocks x nstep x ndraw array of impulse responses.  Returned
value is not printed, but can be assigned if you want to keep it.
}
\description{
Uses output from \code{postdraw} or \code{SVARpostdraw()} to construct impulse
responses with error bands.
}
\details{
Using \code{postdraw()} output as the pdout argument, irf's with triangular orderings
are found if there is no \code{smat} argument.  Using \code{SVARpostdraw()} output, the
\code{smat} array is filled with the inverses of pdout$A.
}
